requirements.txt
torch
transformers
sentence-transformers
faiss-cpu
langgraph
langchain
tiktoken
pandas
numpy
tqdm
sentencepiece
accelerate
streamlit
